Company Information

Provides general SpaceX company information

company_info()

Value

A list of useful information about SpaceX

name

Company name

founder

Founder's name

founded

Year founded

employees

Number of employees

vehicles

Number of main types of launch vehicles

launch_sites

Number of launch sites

ceo

Name of the current CEO

cto

Name of the current CTO

coo

Name of the current COO

cto_propulsion

Name of the current CTO for Propulsion

valuation

Current company valuation (in USD)

headquarters

List of address components for the company headquarters

summary

Brief company description

Examples

company_info()

#> $name
#> [1] "SpaceX"
#> 
#> $founder
#> [1] "Elon Musk"
#> 
#> $founded
#> [1] 2002
#> 
#> $employees
#> [1] 6000
#> 
#> $vehicles
#> [1] 3
#> 
#> $launch_sites
#> [1] 3
#> 
#> $test_sites
#> [1] 1
#> 
#> $ceo
#> [1] "Elon Musk"
#> 
#> $cto
#> [1] "Elon Musk"
#> 
#> $coo
#> [1] "Gwynne Shotwell"
#> 
#> $cto_propulsion
#> [1] "Tom Mueller"
#> 
#> $valuation
#> [1] 1.5e+10
#> 
#> $headquarters
#> $headquarters$address
#> [1] "Rocket Road"
#> 
#> $headquarters$city
#> [1] "Hawthorne"
#> 
#> $headquarters$state
#> [1] "California"
#> 
#> 
#> $summary
#> [1] "SpaceX designs, manufactures and launches advanced rockets and spacecraft. The company was founded in 2002 to revolutionize space technology, with the ultimate goal of enabling people to live on other planets."
#>