Vehicle Information

Provides information and stats on SpaceX company vehicles, both launch vehicles and capsules

vehicle_info(id = NULL)

Arguments

id
If not provided, gives a list of all vehicles. Otherwise, just the specified vehicle.

Value

A list vehicle parameters

Examples

vehicle_info("falcon9")

#> $id
#> [1] "falcon9"
#> 
#> $name
#> [1] "Falcon 9"
#> 
#> $active
#> [1] TRUE
#> 
#> $stages
#> [1] 2
#> 
#> $cost_per_launch
#> [1] 62000000
#> 
#> $success_rate_pct
#> [1] 94
#> 
#> $first_flight
#> [1] "2010-06-04"
#> 
#> $country
#> [1] "United States"
#> 
#> $company
#> [1] "SpaceX"
#> 
#> $height
#> $height$meters
#> [1] 70
#> 
#> $height$feet
#> [1] 229.6
#> 
#> 
#> $diameter
#> $diameter$meters
#> [1] 3.7
#> 
#> $diameter$feet
#> [1] 12
#> 
#> 
#> $mass
#> $mass$kg
#> [1] 549054
#> 
#> $mass$lb
#> [1] 1207920
#> 
#> 
#> $payload_weights
#>     id                          name    kg    lb
#> 1  leo               low earth orbit 22800 50265
#> 2  gto geosynchronous transfer orbit  8300 18300
#> 3 mars                    mars orbit  4020  8860
#> 
#> $first_stage
#> $first_stage$reusable
#> [1] TRUE
#> 
#> $first_stage$engines
#> [1] 9
#> 
#> $first_stage$fuel_amount_tons
#> [1] 385
#> 
#> $first_stage$burn_time_sec
#> [1] 180
#> 
#> $first_stage$thrust_sea_level
#> $first_stage$thrust_sea_level$kN
#> [1] 7607
#> 
#> $first_stage$thrust_sea_level$lbf
#> [1] 1710000
#> 
#> 
#> $first_stage$thrust_vacuum
#> $first_stage$thrust_vacuum$kN
#> [1] 8227
#> 
#> $first_stage$thrust_vacuum$lbf
#> [1] 1849500
#> 
#> 
#> 
#> $second_stage
#> $second_stage$engines
#> [1] 1
#> 
#> $second_stage$fuel_amount_tons
#> [1] 90
#> 
#> $second_stage$burn_time_sec
#> [1] 397
#> 
#> $second_stage$thrust
#> $second_stage$thrust$kN
#> [1] 934
#> 
#> $second_stage$thrust$lbf
#> [1] 210000
#> 
#> 
#> $second_stage$payloads
#> $second_stage$payloads$option_1
#> [1] "dragon"
#> 
#> $second_stage$payloads$option_2
#> [1] "composite fairing"
#> 
#> $second_stage$payloads$composite_fairing
#> $second_stage$payloads$composite_fairing$height
#> $second_stage$payloads$composite_fairing$height$meters
#> [1] 13.1
#> 
#> $second_stage$payloads$composite_fairing$height$feet
#> [1] 43
#> 
#> 
#> $second_stage$payloads$composite_fairing$diameter
#> $second_stage$payloads$composite_fairing$diameter$meters
#> [1] 5.2
#> 
#> $second_stage$payloads$composite_fairing$diameter$feet
#> [1] 17.1
#> 
#> 
#> 
#> 
#> 
#> $engines
#> $engines$number
#> [1] 9
#> 
#> $engines$type
#> [1] "merlin"
#> 
#> $engines$version
#> [1] "1D+"
#> 
#> $engines$layout
#> [1] "octaweb"
#> 
#> $engines$engine_loss_max
#> [1] 2
#> 
#> $engines$propellant_1
#> [1] "liquid oxygen"
#> 
#> $engines$propellant_2
#> [1] "RP-1 kerosene"
#> 
#> $engines$thrust_sea_level
#> $engines$thrust_sea_level$kN
#> [1] 845
#> 
#> $engines$thrust_sea_level$lbf
#> [1] 190000
#> 
#> 
#> $engines$thrust_vacuum
#> $engines$thrust_vacuum$kN
#> [1] 914
#> 
#> $engines$thrust_vacuum$lbf
#> [1] 205500
#> 
#> 
#> $engines$thrust_to_weight
#> [1] 180.1
#> 
#> 
#> $landing_legs
#> $landing_legs$number
#> [1] 4
#> 
#> $landing_legs$material
#> [1] "carbon fiber"
#> 
#> 
#> $description
#> [1] "Falcon 9 is a two-stage rocket designed and manufactured by SpaceX for the reliable and safe transport of satellites and the Dragon spacecraft into orbit."
#>